Minimal handling of implementation logs (messages).
