+++ pyEDAA/OutputFilter/Xilinx/Synthesis.py
@@ -29,85 +29,17 @@
 # ==================================================================================================================== #
 #
 """A filtering anc classification processor for AMD/Xilinx Vivado Synthesis outputs."""
-from datetime import datetime
-from enum     import Flag
 from pathlib  import Path
-from re       import compile as re_compile, Pattern
-from typing   import ClassVar, List, Optional as Nullable, Callable, Dict, Type, Generator, Iterator, TypedDict, Union
+from re       import compile as re_compile
+from typing   import ClassVar, List, Callable, Dict, Type, Iterator, Union
 
-from pyTooling.Decorators import export, readonly
-from pyTooling.MetaClasses import ExtendedType, abstractmethod, mustoverride
+from pyTooling.Decorators  import export, readonly
+from pyTooling.MetaClasses import mustoverride
 from pyTooling.Common      import firstValue
-from pyTooling.Stopwatch import Stopwatch
-from pyTooling.Versioning  import YearReleaseVersion
 
-from pyEDAA.OutputFilter.Xilinx import VivadoMessage, VivadoInfoMessage, VivadoWarningMessage, VivadoCriticalWarningMessage, VivadoErrorMessage
-
-
-@export
-class ProcessingState(Flag):
-	Processed =        1
-	Skipped =          2
-	EmptyLine =        4
-	CommentLine =      8
-	DelimiterLine =   16
-	TableLine =       32
-	TableHeader =     64
-	Reprocess =      512
-	Last =          1024
-
+from pyEDAA.OutputFilter.Xilinx import VivadoMessage, ProcessingState, Parser, Preamble, BaseProcessor
 
 TIME_MEMORY_PATTERN = re_compile(r"""Time \(s\): cpu = (\d{2}:\d{2}:\d{2}) ; elapsed = (\d{2}:\d{2}:\d{2}) . Memory \(MB\): peak = (\d+\.\d+) ; gain = (\d+\.\d+)""")
-
-@export
-class Parser(metaclass=ExtendedType, slots=True):
-	_processor: "Processor"
-
-	def __init__(self, processor: "Processor"):
-		self._processor = processor
-
-	@readonly
-	def Processor(self) -> "Processor":
-		return self._processor
-
-	@abstractmethod
-	def ParseLine(self, lineNumber: int, line: str) -> ProcessingState:
-		pass
-
-
-@export
-class Preamble(Parser):
-	_toolVersion:   Nullable[YearReleaseVersion]
-	_startDatetime: Nullable[datetime]
-
-	_VERSION:   ClassVar[Pattern] = re_compile(r"""# Vivado v(\d+\.\d(\.\d)?) \(64-bit\)""")
-	_STARTTIME: ClassVar[Pattern] = re_compile(r"""# Start of session at: Thu (\w+) (\d+) (\d+):(\d+):(\d+) (\d+)""")
-
-	def __init__(self, processor: "Processor"):
-		super().__init__(processor)
-
-		self._toolVersion =   None
-		self._startDatetime = None
-
-	@readonly
-	def ToolVersion(self) -> YearReleaseVersion:
-		return self._toolVersion
-
-	@readonly
-	def StartDatetime(self) -> datetime:
-		return self._startDatetime
-
-	def ParseLine(self, lineNumber: int, line: str) -> ProcessingState:
-		if self._toolVersion is not None and line.startswith("#----"):
-			return ProcessingState.DelimiterLine | ProcessingState.Last
-		elif (match := self._VERSION.match(line)) is not None:
-			self._toolVersion = YearReleaseVersion.Parse(match[1])
-			return ProcessingState.Processed
-		elif (match := self._VERSION.match(line)) is not None:
-			self._startDatetime = datetime(int(match[6]), int(match[1]), int(match[2]), int(match[3]), int(match[4]), int(match[5]))
-			return ProcessingState.Processed
-
-		return ProcessingState.Skipped
 
 
 @export
@@ -385,6 +317,7 @@ class WritingSynthesisReport(Section):
 
 		return super().ParseLine(lineNumber, line)
 
+
 PARSERS = (
 	Preamble,
 	RTLElaboration,
@@ -408,68 +341,26 @@ PARSERS = (
 	WritingSynthesisReport,
 )
 
+
 Parsers = Union[*PARSERS]
 
+
 @export
-class Processor(metaclass=ExtendedType, slots=True):
+class Processor(BaseProcessor):
 	_logfile:  Path
 	_parsers:  Dict[Type[Parser], Parsers]
 	_state:    Callable[[int, str], bool]
-	_duration: float
 
-	_infoMessages:            List[VivadoInfoMessage]
-	_warningMessages:         List[VivadoWarningMessage]
-	_criticalWarningMessages: List[VivadoCriticalWarningMessage]
-	_errorMessages:           List[VivadoErrorMessage]
-	_toolIDs:                 Dict[int, str]
-	_toolNames:               Dict[str, int]
-	_messagesByID:            Dict[int, Dict[int, List[VivadoMessage]]]
+	_activeParsers: List[Parsers]
 
 	def __init__(self, synthesisLogfile: Path):
+		super().__init__()
+
 		self._logfile =  synthesisLogfile
 		self._parsers =  {p: p(self) for p in PARSERS}
 		self._state =    firstValue(self._parsers).ParseLine
-		self._duration = 0.0
 
-		self._infoMessages =            []
-		self._warningMessages =         []
-		self._criticalWarningMessages = []
-		self._errorMessages =           []
-		self._toolIDs =                 {}
-		self._toolNames =               {}
-		self._messagesByID =            {}
-
-	@readonly
-	def Duration(self) -> float:
-		return self._duration
-
-	@readonly
-	def ToolIDs(self) -> Dict[int, str]:
-		return self._toolIDs
-
-	@readonly
-	def ToolNames(self) -> Dict[str, int]:
-		return self._toolNames
-
-	@readonly
-	def MessagesByID(self) -> Dict[int, Dict[int, List[VivadoMessage]]]:
-		return self._messagesByID
-
-	@readonly
-	def InfoMessages(self) -> List[VivadoInfoMessage]:
-		return self._infoMessages
-
-	@readonly
-	def WarningMessages(self) -> List[VivadoWarningMessage]:
-		return self._warningMessages
-
-	@readonly
-	def CriticalWarningMessages(self) -> List[VivadoCriticalWarningMessage]:
-		return self._criticalWarningMessages
-
-	@readonly
-	def ErrorMessages(self) -> List[VivadoErrorMessage]:
-		return self._errorMessages
+		self._activeParsers = list(self._parsers.values())
 
 	@readonly
 	def HasLatches(self) -> bool:
@@ -489,9 +380,6 @@ class Processor(metaclass=ExtendedType, slots=True):
 		except KeyError:
 			yield from ()
 
-	def __getitem__(self, item: Type[Parser]) -> Parsers:
-		return self._parsers[item]
-
 	@readonly
 	def HasBlackboxes(self) -> bool:
 		return len(self._parsers[WritingSynthesisReport]._blackboxes) > 0
@@ -500,111 +388,29 @@ class Processor(metaclass=ExtendedType, slots=True):
 	def Cells(self) -> Dict[str, int]:
 		return self._parsers[WritingSynthesisReport]._cells
 
-	def Parse(self):
-		with Stopwatch() as sw:
-			with self._logfile.open("r", encoding="utf-8") as f:
-				content = f.read()
+	def __getitem__(self, item: Type[Parser]) -> Parsers:
+		return self._parsers[item]
 
-			activeParsers = list(self._parsers.values())
+	def _Parse(self, lineNumber: int, line: str):
+		if self._state is not None:
+			result = self._state(lineNumber, line)
+			if ProcessingState.Last in result:
+				obj: Section = self._state.__self__
+				self._activeParsers.remove(obj)
+				self._state = None
 
-			lines = content.splitlines()
-			for lineNumber, line in enumerate(l.rstrip() for l in lines):
-				prefix = line[:4]
-				if prefix == "INFO":
-					if (message := VivadoInfoMessage.Parse(line, lineNumber)) is None:
-						print(f"pattern not detected\n{line}")
-						continue
-
-					self._infoMessages.append(message)
-				elif prefix == "WARN":
-					if (message := VivadoWarningMessage.Parse(line, lineNumber)) is None:
-						print(f"pattern not detected\n{line}")
-						continue
-
-					self._warningMessages.append(message)
-				elif prefix == "CRIT":
-					if (message := VivadoCriticalWarningMessage.Parse(line, lineNumber)) is None:
-						print(f"pattern not detected\n{line}")
-						continue
-
-					self._criticalWarningMessages.append(message)
-				elif prefix == "ERRO":
-					if (message := VivadoErrorMessage.Parse(line, lineNumber)) is None:
-						print(f"pattern not detected\n{line}")
-						continue
-
-					self._errorMessages.append(message)
+				print(f" DONE: {obj.__class__.__name__}")
+		else:
+			if line.startswith("Start") or line.startswith("Starting"):
+				for parser in self._activeParsers:   # type: Section
+					if line.startswith(parser._START):
+						print(f"BEGIN: {parser.__class__.__name__}")
+						self._state = parser.ParseLine
+						_ = self._state(lineNumber, line)
+						break
 				else:
-					if self._state is not None:
-						result = self._state(lineNumber, line)
-						if ProcessingState.Last in result:
-							obj: Section = self._state.__self__
-							activeParsers.remove(obj)
-							self._state = None
+					print(f"Unknown section\n  {line}")
 
-							print(f" DONE: {obj.__class__.__name__}")
-					else:
-						if line.startswith("Start") or line.startswith("Starting"):
-							for parser in activeParsers:   # type: Section
-								if line.startswith(parser._START):
-									print(f"BEGIN: {parser.__class__.__name__}")
-									self._state = parser.ParseLine
-									_ = self._state(lineNumber, line)
-									break
-							else:
-								print(f"Unknown section\n  {line}")
-
-					continue
-
-				if message._toolID in self._messagesByID:
-					sub = self._messagesByID[message._toolID]
-					if message._messageKindID in sub:
-						sub[message._messageKindID].append(message)
-					else:
-						sub[message._messageKindID] = [message]
-				else:
-					self._toolIDs[message._toolID] = message._toolName
-					self._toolNames[message._toolName] = message._toolID
-					self._messagesByID[message._toolID] = {message._messageKindID: [message]}
-
-		self._duration = sw.Duration
-
-
-def main():
-	logfile = Path("tests/data/Stopwatch/toplevel.vds")
-	logfile = Path("tests/data/ADL-1000/toplevel.vds")
-	processor = Processor(logfile)
-	processor.Parse()
-
-	print()
-	print(f"Vivado version: {processor._parsers[Preamble]._toolVersion}")
-
-
-	#
-	# print("%" * 80)
-	# for toolID, messageGroups in processor._messagesByID.items():
-	# 	print(f"{toolID} ({len(messageGroups)}):")
-	# 	for messageID, messages in messageGroups.items():
-	# 		print(f"  {messageID} ({len(messages)}):")
-	# 		for message in messages:
-	# 			print(f"    {message}")
-	#
-	# print("%" * 80)
-	# for tool, toolID in processor._toolNames.items():
-	# 	messages = list(chain(*processor._messagesByID[toolID].values()))
-	# 	print(f"{tool} ({len(messages)}):")
-	# 	for message in messages:
-	# 		print(f"  {message}")
-
-	# print(f"Cells Statistics")
-	# for cellName, cellCount in processor._parsers[WritingSynthesisReport]._cells.items():
-	# 	print(f"  {cellName:16}: {cellCount}")
-
-if __name__ == '__main__':
-	main()
-
-
-# latches
 # unused
 # used but not set
 # statemachine encodings
