+++ tests/unit/Vivado/SynthesisLog.py
@@ -32,9 +32,11 @@
 from pathlib  import Path
 from unittest import TestCase as TestCase
 
-from pyTooling.Versioning import YearReleaseVersion
+from pyTooling.Versioning       import YearReleaseVersion
 
-from pyEDAA.OutputFilter.Xilinx.Synthesis import Processor, Preamble, WritingSynthesisReport
+from pyEDAA.OutputFilter.Xilinx                import Preamble
+from pyEDAA.OutputFilter.Xilinx.Synthesis      import Processor as SynthProc, WritingSynthesisReport
+from pyEDAA.OutputFilter.Xilinx.Implementation import Processor as ImplProc
 
 if __name__ == "__main__": # pragma: no cover
 	print("ERROR: you called a testcase declaration file as an executable module.")
@@ -45,7 +47,7 @@ if __name__ == "__main__": # pragma: no cover
 class Stopwatch(TestCase):
 	def test_SynthesisLogfile(self) -> None:
 		logfile = Path("tests/data/Stopwatch/toplevel.vds")
-		processor = Processor(logfile)
+		processor = SynthProc(logfile)
 		processor.Parse()
 
 		self.assertLess(processor.Duration, 0.1)
@@ -59,3 +61,17 @@ class Stopwatch(TestCase):
 
 
 		self.assertEqual(0, len(processor[WritingSynthesisReport].Blackboxes))
+
+	def test_ImplementationLogfile(self) -> None:
+		logfile = Path("tests/data/Stopwatch/toplevel.vdi")
+		processor = ImplProc(logfile)
+		processor.Parse()
+
+		self.assertLess(processor.Duration, 0.1)
+
+		self.assertEqual(152, len(processor.InfoMessages))
+		self.assertEqual(2, len(processor.WarningMessages))
+		self.assertEqual(2, len(processor.CriticalWarningMessages))
+		self.assertEqual(0, len(processor.ErrorMessages))
+
+		self.assertEqual(YearReleaseVersion(2025, 1), processor[Preamble].ToolVersion)
